ceph: Switch to central ceph-ansible repository
All the ceph roles are generated from the central ceph-ansible
repository. However, it appears that the script (or CI) that
generates these roles has been broken for a while. The last update
for generated roles was nearly 2 months ago. This means that we
are missing a lot of ceph features and fixes that exist in the central
repo but not on the generated roles. Lets switch to the main
ceph-ansible repository for our Ceph deployments in order to ensure
that we are always using the latest code.

Change-Id: Ic907ee734a95e8af7c4c51568cc0870607b7b2f9