Reorder cinder service installs
The recommended order that services restart during an upgrade of cinder
is scheduler, volume, backup, and then api. Update the playbook so that
cinder services are also installed in that same order.

Also add back the reloading of services other than cinder-api, as the
misordering may have been the cause of high CPU usage in the past.

Depends-On: I6f39b4da6d4d8a3768f157c73970fe0177885ff3

Change-Id: I81874ffea0afd499228698422842da252fce7dc0