Add a return code check to the osa wrapper
Due to our use of block/rescue pattern in some of the playbooks/roles
the after run report will show something "failed" even when the run
succeeds. This creates a poor user experience when looking at the
playbook runs, especially as a new user. This change adds a simple
return code check when the openstack-ansible wrapper is used and reports
if the run succeeds or is a failure based on the RC. This should help
users understand when something actually fails, at a glance.

> NOTE: This change should no longer be needed as soon as we're able
  to move to ansible 2.5 which improves the UX when the block/rescue
  pattern is used.

With this change a user will see an "EXIT NOTICE" which is reported in a
way that resembles how ansible reports information on a playbook run
which should be simple to see and understand in almost all cases.

Change-Id: Ic54a7524c09a170e20830c5f8d2c2a0658159ed0
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit f0d8ffd5c147fe07335b55ee0f04785a22871e5c)