Stable keepalived priorities
The calculation of the keepalived priority depended on the set of hosts
in the ansible_play_hosts variable. This meant that priorites changed
when the playbook was limited to only run on a subset of the hosts in
the "haproxy" group (eg. when reinstalling just one node).

This change replaces ansible_play_hosts by groups["haproxy"] so that the
priority calculation is stable as long as the set of haproxy nodes does
not change. It also simplifies the priority calculation without
changeing the result.

This change does not change the configuration in case the playbook is
run without --limit and ensures that the configuration does not change
if it is run with --limit.

To make sure the group vars resolution is not done in a all group,
which might have 0 haproxy nodes, we ensure the group vars are applied
properly to the haproxy group. This is a partial backport of the
patch to limit group_vars scoping [1].

[1]: https://github.com/openstack/openstack-ansible/commit/85501cbf263029cfd5be44b9274113e4e5e4b6f8#diff-849f58a4930e8d8f73fd8310002552a9

Co-Authored-By: Jean-Philippe Evrard <jean-philippe@evrard.me>
Change-Id: I378351954fbed7b1fc31303c60e4575ddad9cf1d
(cherry picked from commit 78ed74957dfbd1bb151d0e6b5187fa0c263881ad)