Move doc requirements to own file
OpenStack-Infra now allows the documentation requirements
to be held in their own file instead of being mixed in with
the other requirements. During the changes to enable this,
our previous method of implementing sphinxmark no longer
works due to the jobs no longer using tox at all for the
tests.

We cannot have sphinxmark in newton's test-requirements.txt
as it is not a library included in the global-requirements
file (it is from ocata onwards), so here we move it to the
docs requirements file.

Change-Id: I802f1f56d1a9ecd41fbc9bcec52f5fd7827e38fb