Correctly set nova_cinder_rbd_inuse in os_nova execution
This line is already present in Pike and master, however it was
omitted by mistake in Ocata. This means that when Cinder uses RBD
backends, but no RBD backend is used for Nova's instance backing,
the ceph_client role will be skipped when deploying the compute host.

When this happens, instances trying to attach volumes will receive
an error stating that the libvirt secret store does not contain the
RBD secret:
"libvirtError: Secret not found: no secret with matching uuid"

By passing this nova_cinder_rbd_inuse boolean to the Nova role based
on the set_fact above in this playbook, ceph_client will deploy the
ceph libraries and secret to the compute host when it is needed
for volume attachments to work.

Change-Id: I4a673c4966e36ccc88392ff9e103da7806248e3c