Cleanup the gate-check-commit dstat part
Following the advice given in review [1], we could put the
conditional inside the scripts-library to simplify the behavior
inside the gate-check-commit script.

On top of that, if we pipe to || true, the function wouldn't
fail on branches where the function/vars are not defined,
cleaning the code further.

[1]: https://review.openstack.org/#/c/531131/2

Change-Id: Icda339747267cf6bf514d45878c26533b5341f83
(cherry picked from commit 7f192b104f3c27fddfc2f86b9b7d61e9222b124c)