Fix keepalived warnings
Relative paths for keepalived scripts cause warnings to appear in
the systemd journal:

    Unable to access script `kill`
    Disabling track script haproxy_check_script since not found
    WARNING - script `ping` resolved by path search to
      `/usr/bin/ping`. Please specify full path.

This patch ensures that full paths are used for both of the current
keepalived scripts and it removes the warnings from the logs.

It also ensures that script security is enabled. This prevents
keepalived from running a script as root that a normal user can
write to.

Closes-Bug: 1742487
Change-Id: Ie3bc334a1669c976298d7ef1eb7361e85fade7e1