Fix logic to check for insecure Keystone
The logic to check for insecure Keystone would only validate if
both HAproxy and Keystone SSL is installed.  However, in the case
where SSL happens at either of those (ex. SSL at HAproxy only),
then it would still declare it as insecure (it is not however
as we're hitting the VIP which does SSL termination).

This changes the behaviour to check if the protocol is HTTPs and
either the Keystone or HAproxy certificates are defined then it
considers it as secure, otherwise it is insecure.

Closes-Bug: #1745479
Co-Authored-By: Logan V <logan2211@gmail.com>
Change-Id: I3683c5d50cab2d321337c71bed565d2e2ab861fc