Vagrantfile: Use rsync to copy working directory to VM
Depending on the VM distro the working directory can either be rsync'd
to the VM or mounted. When the directory is mounted, various artifacts
are left behind from the test execution which is not desired especially
when the directory is shared between multiple VMs for parallel testing.
As such, it's best to simply copy the directory to the VM using 'rsync'
so each VM can get its own copy.

Change-Id: I2ef84692b57e136fc6008f33f3e48f861f9a58da
Link: I1aab7d3f0d26f35dea84ae088fe78cfd2de79366