Remove impossible tests
We haven't managed to pass functional tests for CentOS this cycle,
it seems overly ambitious to think the upgrade tests would pass.

We also removed the periodics from centos/suse, because they are
all timing out, and therefore are using resources for nothing.
We keep the check jobs to be aware of timings.

Change-Id: I4900cbc22da83582ddd48af7df99d66de1a129f9