bootstrap-host: Create swap without using a failed task
This reverts commit 541f04e54683547182580ced58adfcdb2c920043 and
implements a more targeted solution.

Unfortunately this implementation's of a try/rescue system which
will always fail the try on a fresh system causes a lot of confusion
for anyone new to OSA. When seeing the reported failed task the
user thinks something went wrong, even though it was working as
designed.

The main.yml uses a conditional include for the execution of these
tasks only if the current swap available < 1MB. This should be enough
to prevent the tasks executing again. If they do execute again, then
the task to enable swap is now fully idempotent too, so the reason
for the initial implementation is also solved.

Change-Id: I7be78780505f14fb1b581d22d4762a34cd2005d8