Run openstack_openrc before Magnum installation
The Magnum role attempts to upload images for the operation of
Magnum, however, the role for openrc is installed after magnum
which means the role cannot use the 'default' cloud profile
to upload images.

This patch adjusts the order so that the os_magnum role is able
to upload images.

Change-Id: I10846e5844c2f02c1a5450eaa950b3704672c324