Start nova-conductor container first
On stable/newton, if you reboot an AIO the nova-conductor and
nova-api-metadata services start at the same time, and it appears that
nova-conductor needs to be up first before nova-api-metadata starts.
When this isn't the case, the nova-api-metadata service eventually
times out (cleanly), and requires a manual start/restart to bring it
online.

This commit simply gives the nova-conductor containers a start order of
39, and the remaining nova containers a start order of 49. This ensures
that nova-conductor starts before nova-api-metadata comes up.

Change-Id: I5fe06ee86070c8e30fa3e169a128e95b660df4df