Fix LXC volume size
With the change to allow volumes to dynamically grow we need to define
the default volume size to contain a valid size modifier. At present the
value is creating a sparse file that is 1024 bytes. This change sets the
size to gigabytes.

The role pin is updated to include the changes in the lxc_hosts role
which use this method of configuration. Without it the quota setting
end up with a GG at the end, which is invalid and the command fails.

Change-Id: Ib9b23b023d2ad03ce72c4abce5dc9e51d4643ac0
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>