Update os_magnum role pin
The magnum source repository unfortunately is not
adhering to the global requirements management process
in that it has requirements specified in the repo which
are not in the global requirements repo.

As a result of the requirement being missing, those
requirements are not constrained by upper-constraints.
That, in turn, causes the repo build to fail to build
the magnum venv due to incompatible requirements.

This patch brings in the updated role SHA which pins
kubernetes and related dependencies and the release note
to communicate the change.

Change-Id: I49312a568ff6ce74497f00eec288166707e33fc0