Have zuul check out ansible for devel AIO job
Zuul is configured to understand ansible/ansible on github. Add ansible
as a required-project for the devel job so that depends-on patches will
work. Once we're happy with this, we can add the devel-aio job to be run
on PRs to os_ modules.

Change-Id: I12a74d97593c8780047d3dae9590f4e1fde3859b
(cherry picked from commit 148cc97d61b74b73cca8c799250f053c23ec853e)