Ensure that constraints are used for AIO bootstrap
This is a partial backport of [1] without the octavia
specific changes, and [2] combined with it.

Currently the pip_install role does not configure any
constraints when exercising the AIO bootstrap, creating
the possibility that the initial bootstrap may not have
the right software versions.

This patch ensures that the same constraints are used
for the AIO bootstrap as was done for the Ansible
bootstrap.

[1] https://review.openstack.org/496929
[2] https://review.openstack.org/545690
Change-Id: I6aa467d4f1f6db33c824f4fcf8e46adb0a26cfa0