Replace host_need_pip with standard override mechanism
The variable host_need_pip is currently used to be selective
about when pip is installed on the target hosts/containers.

This is an unnecessary precaution given that if the repo
container is not yet up, there is already a mechanism in place
to detect this and override the constraint and get-pip script
URL. By using this mechanism we get pip installed very early
and pip's configuration is still correctly constrained.

This patch ensures that this very same mechanism is used across
the board instead of using two different mechanisms, thereby
simplifying the approach.

Change-Id: I22d2f5edf640b71cb80b3fd1de597ea7391db11c
Implements: blueprint python-build-install-simplification
Signed-off-by: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>