Increase Ansible SSH connection retries
Retries for Ansible SSH connections have been enabled in master but not
the stable branches. Testing using openstack-ansible-ops multi-node AIOs
results in failures due to SSH connection errors on a regular basis.

This change increases the number of retries from 3 to 5 to increase the
time allowed for a connection to be made from approximately 20 seconds
to approximately 50 seconds. This will then provide for this setting to
be backported to all stable branches.

Change-Id: I45dcf6313e1daf63b35b8620b3467eebac0eb95d