CentOS 7 integrated gate optimization
The CentOS 7 AIO jobs time out fairly frequently in the gates and
this patch switches the jobs to use the new aio_basekit scenario in
the hopes that jobs will complete more often.

This patch also sets SELinux to Permissive mode for CentOS 7 gate
testing. This can be changed back to Enforcing when CentOS 7's
SELinux policies are ready.

This patch also sets package_state to present for CentOS 7 jobs. This
is a temporary measure to ensure more of the gate jobs can finish on
time and we can expose deployment issues in the gates.

(This backport fixes a cherry-pick conflict in project.yaml.)

Change-Id: Ic478a7402fee3df0667f0d6fd9a714688808b2ec
(cherry picked from commit f4bc81cff97e943a28d89213305e80f5a5ccac79)