Isolate the Ansible bootstrap
As we move forward with the python build/install simplification,
more of the config is being applied in the pip conf, rather than
being given as CLI options in each ansible task. This is good
for providing a consistent experience, but has side effects for
things like the Ansible bootstrap.

When bootstrapping Ansible, we want to make use of a specific
set of constraints which is not polluted by additional config
on the host. We do not want the constraints from the repo
server because they happen to include a constraint for Ansible
which clashes with our requirement. As such, we only want the
native OpenStack upper-constraints.

In this patch we use the --isolated CLI option to ignore the
host's pip configuration. We also introduce the ability to set
extra pip install options by using the PIP_OPTS environment
variable. This can be used to set the installation to use an
alternative pypi mirror, or for extra links, etc.

Change-Id: Ic966bafd04c4c01b3d93851a0e3ec2c1f3312f28
Closes-Bug: #1751316