Remove non used variables
Remove the proto, port and url_v2 variables and set only the
needed and used internaluri variable.
This is part of a cleanup that we started at ptg to remove
variables that present only to create other variables.
Such variables are not needed and should be removed to conserve memory.

Depends-On: I6fa28dcdb251da4592e10b288d15a208e83b097d
Depends-On: I891b94a8fb452e9fb5166e7c322d152060efdcce
Change-Id: Ie74669fa1660e11acb0dec481a3a01fdc066bc27