Avoid putting rgw conf on ALL hosts
To avoid adding unnecessary RGW configuration to hosts that don't
require it we should separate out the rgw_overrides to only apply to
ceph.conf on RGW hosts.

Additionally, the radosgw_dns_name var is no longer used by ceph-ansible
or OSA in any way, so this has been cleaned up and removed.

Change-Id: I243e742da3bf7d1e452998f257f7d0804e13237b