Replace meta job map for aio->aio_lxc with simpler mechanism
The job map does not appear to be working as it was designed
to as the pike->queens upgrades are failing due to an incorrect
scenario name being given to the pike bootstrap process.

The patch also correctly sets the scenario for the jobs that
run, and also replaces the mapping with a simpler mechanism in
the gate-check-commit script.

Change-Id: I384df79df83b04adfd7ba002ce54b9044c6df5e1