Pin ansible at urllib fix SHA
In ansible 2.3.2.0 there is an SNI bug which is preventing deployment
in calico environments since get_url is unable to fetch packages
from github.com.

Example output of the issue:
fatal: [aio1]: FAILED! => {"attempts": 5, "changed": false, "failed": true,
"msg": "Failed to validate the SSL certificate for github.com:443. Make sure
your managed systems have a valid CA certificate installed. You can use
validate_certs=False if you do not need to confirm the servers identity but
this is unsafe and not recommended. Paths checked for this platform:
/etc/ssl/certs, /etc/pki/ca-trust/extracted/pem, /etc/pki/tls/certs,
/usr/share/ca-certificates/cacert.org, /etc/ansible. The exception msg
was: (\"bad handshake: Error([('SSL routines', 'ssl3_read_bytes',
'tlsv1 alert protocol version')],)\",)."}

Pinning to the SHA [1] fixes this issue, and it will be included in
the future 2.3.4.0 release of ansible (currently in rc1). We can
change this pin to use the 2.3.4.0 release once it is final.

[1] https://github.com/ansible/ansible/commit/6020c0d4f64dbafcbafdffbfc67cfc5bb6b3c75c

Change-Id: I6f7bd8078021fb2dd2ac1978d4511c12b76d318a