Move CentOS to experimental in Ocata
We've been overly enthusiastic with CentOS in the past, and
Ocata is very unreliable. Right now, there are lots of times
where it runs in periodics and fails. Because nobody can
take the time to ensure ocata works with CentOS, we should remove
it from the periodics, and add it as experimental jobs for that
cycle.

In the future, it can be changed back, if people work on its
reliability under the Ocata branch, like it's done for the
Rocky and Queens branches.

Change-Id: Ia622045212452a20ad9ce8a2e0def1233f5839f1