[Docs] Restructure inventory documentation
The documentation of the inventory started to be spread out, but
also massive to be inside one page.

This moves it to the inventory section of the reference, but at the
same time, to improve readability, moves the previous content into
sub pages.

Change-Id: If2c6b83abafdc66879d818df4c9690142389a965
(cherry picked from commit 181c83baef8aa87ef2374cef961eddf3d036f2ad)