bootstrap-host: Prepare disk for machinectl storage
Currently we prepare an available secondary disk, but do not
use it. The preparation taking place now was useful when the
backing file system for LXC was 'dir', but it is now machinectl
which uses a btrfs partition instead.

This patch implements the preparation of a btrfs sparse file
if there is no secondary disk, or the preparation of a secondary
btrfs disk if there is one.

The user_variables override for lxc_host_machine_volume_size
is removed as the default value in the role is perfectly adequate
for AIO deployment needs.

We also add some diagnostic information to help troubleshoot
issues if they arise.

Change-Id: Icadbcfd420bbe7c14a04606790cab09a1dce03ce