common-tasks: Container Start/Stop conditions
The container start/stop conditions for lxc can fail if a container is
slow to start/stop which may result in an retry and fail due to the
confused state. This change checks the stdout of the command which will
report if the container is already started or stopped which should
ensure we're not failing a playbook run simply because of a container
is slow to respond to a command or in an otherwise confused state.

This issue is often seen on the cent gates:
http://logs.openstack.org/47/552047/23/check/openstack-ansible-deploy-aio_basekit-centos-7/222e1b2/job-output.txt.gz#_2018-03-15_01_45_14_559041

Change-Id: Ic52c1d00592fb194889a3aef695075c3e102fc4c
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>