[Docs] Fix doc building idempotency
When running the doc building a new scenario was generated,
and appened to the scenario-table-gen.html page. We should instead
generate that page from scratch, and ensure it's always the
same content. This doesn't clean the scenario table gen, it simply
makes sure the generated content is always ordered the same way.

Change-Id: I9098f5d76b73be5ef28684027ff5039cd52d6ddb