Ensure package cache is updated appropriately
When the repo container is removed, the apt cache config
is still in place and causes problems. This is due to the
tasks running to remove the cache configuration not
executing early enough in the deployment.

This patch re-orders the tasks so that the removal can
be done and the cache is updated afterwards to confirm
a working state.

It also makes the configuration happen on all hosts, then
all containers, at the right timings to ensure that if the
container is removed, the proxy config will also be removed
automatically.

Conflicts:
- playbooks/containers-nspawn-create.yml

Change-Id: I0d1bf9617594d9a5c7c2ea3335c80cc2478df985
(cherry picked from commit 1d731cdd89b660983e5c55e7798dee21557e8fbd)