repo-use.yml: Scope host groups properly
When using multi-region swift or when including more
thank just OpenStack-related hosts in the inventory,
it's important to scope where these tasks are run
properly to prevent attempts at doing things on the
wrong hosts.

The sope of the play now runs against the same scope
for hosts as the openstack-hosts playbook. This is
more consistent and will help prevent unwanted side-
effects.

Change-Id: Ia755845ac2b04f080db45905263059b92072d194