Use production like interfaces in the gate
The gate presently sets up a pseudo networking environment using a
collection of bridges without any interfaces within them. This change
implements the production example interface file using dummy
interfaces. With this change we'll have the ability to test a production
networking stack with bonded interfaces, VLAN tags, and
bridges. To ensure we're enabling this setup in a uniform way the
systemd-networkd role is used to create everything needed.

> Terminal output from what will be created with this change
  http://paste.openstack.org/show/702382/

Change-Id: I7fb15ce81056cf7ad5139333d59787e610bd9c75
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>