Do not collect physical host facts in playbook
The role itself depends on physical host facts being collected,
but the collection occurring in the playbook does not accommodate
--limit containername, it will require
--limit containername:physicalhost in order for the required facts
to be collected.

To allow limit to target only the containers being created, the
fact collection has been moved to the role and the module is
executed against the physical_host using delegation.

Depends-On: https://review.openstack.org/#/c/554105/
Change-Id: Idc0ae7913b3ea9374caa41a207fab8c72f42bf41