[Docs] Include OpenStack-Ansible Manifesto
The OpenStack Ansible Deployment manifesto [1] was lost in times in
pages long forgotten of the wiki.

Instead, it should be part of the documentation, in our reference
section. The reference is a great place because it's where we
explain WHY we are doing things.

To increase visibility of this page, the design decisions are
directly linked from the deploy guide appendix about
"Why OpenStack-Ansible".

Change-Id: Ifb3ef041c773ee6dbe6606e2e5efc0542776dfc5
(cherry picked from commit d99c1d34694bdadf4ce462252e558cbf35164995)