Convert rsyslog to an include_task
The rsyslog role has served us well however there's now a better way
given the ability to remotely journal. This change disables the use of
the `rsyslog_client` and `server` roles unless journal remote is
unavailable. In this change the `rsyslog_client` interaction has been
moved into an include_tasks which is conditionally loaded when the
variable `rsyslog_client_enabled` is set to true. Additionally the
`rsyslog_server` role is disabled unless `rsyslog_server_enabled`
is set true. Using the new variables legacy functionality can be
enabled. By disabling the general rsyslog roles we'll lessen the
overall IO on the cloud and improve the speed of the deployment.

> NOTE: At this time there's no suitable package to install
        "systemd-journal-remote" on opensuse so a conditional has been
        added to ensure distributed log syncing remains functional on
        all of our supported distros. Should a package be made
        available for journal remote we can globally disable the
        legacy rsyslog roles entirely by default.

Change-Id: Ice21667c6999d0ac86b2d7bde648a0375f890210
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>