Update cinder role SHA for cinder_backends fix
Currently the backend setup executes on the first cinder-api
host. The playbook to execute the setup of the API is done
in serial. This creates a bit of a problem on a greenfield
deployment which uses haproxy because the backend is disabled
and there are no other API hosts setup yet. This makes the
CLI commands which require the API to be working fail.

Given that this is a critical bug, this patch updates the
cinder role to ensure that the fix for this is included in
the next release

Change-Id: I23b966c68b491e51f33c89a1aa196a50c42fe3f4
Closes-Bug: #1753125