Apply haproxy vars to haproxy group instead of haproxy_all
Any ansible command loads the group vars/host vars.
If haproxy group isn't properly loaded with hosts, the whole
group_vars/haproxy_all/keepalived.yml will fail to get an index
for any host belonging to the group.

Instead the index should search in the group in which the variable
applies to, to ensure it's always the same kind of group.

In a later commit, we should make sure the index is always done when
group size is bigger than 0.

Change-Id: Id489b3f337ea0db0be0008477c9e3afde3872004
Closes-Bug: #1752932