Add missing service URLs for AODH
The service URLs for internal and admin were not being properly
configured so they were using the default values which means
that any service which was making calls on the internal or
admin URL endpoints would fail.

This patch adds them in order to make them accessible and
have the correct configuration (SSL or no SSL).

Change-Id: I851893c005fc4c91a7aa9a9a979ec315e1fc500f
Closes-Bug: #1750236