Bootstrap lxc_net mtu for gate
Currently the gate assumes the upstream interface in the testing
machine is always running mtu 1500. We change this assumption by
using the facts collected from the host during bootstrap-host to
properly set MTU in the AIO user variables.

This is needed to prevent test failures while downloading internet
artifacts to containers during gate runs on nodepool providers
(such as recently added Limestone) which use a <1500 MTU.

Change-Id: Ifac591c2848c033ad4dfb473bce05860df2bb213