Developer docs refactor
This is a combined cherry-pick from:
- https://review.openstack.org/#/c/557650/
- https://review.openstack.org/#/c/559027/

This refactor has the intent of guiding a new OpenStack-Ansible
user towards his first contributions.

The deploy guide is now pointing towards the Developer Guide,
but there is was no guidance given, just a series of topics.

Here, the first page of the developer guide is an onboarding
page, explaining our repos, what's inside them, and how to
contribute. It then links towards all the sections, which
had to be restructured to match a wizard-style flow.

Change-Id: Ib2a5f2063c4dc58fc249ac94c3cd85bb281f28a5
(cherry picked from commit 3a4ca09d11830ce51c79a04b94170604d73b4eda)