Update all SHAs for newton-eol
This patch updates all the roles to their EOL tag,
copies the release notes from the updated roles into the
integrated repo, and mark the version of this repo as EOL.

The linters check has been disabled from the gates because
it fails in gates while it doesn't fail on a development
machine. It's too late now to change any code now that
the roles are EOL anyway.

Change-Id: I885a0b645a18267b3bbe133d87bca9ed2e589312