Ensure that RDO/SUSE series repo instructions stay up to date
The deployment host documentation gets out of date really quickly.
This patch makes it self maintaining.

Note that due to the fact that this is a stable branch and that
the documentation configuration is vastly different to master,
this is a re-implementation rather than a backport.

Some formatting adjustments have made to prevent the interpretation
of URL's as links by sphinx and to remove the use of multi-line shell
commands.

Closes-Bug: #1759721
Change-Id: Id28540d335390a52e0eb65a734a1d594235265d6