Remove tempest_git_* overrides
Presently, tempest in the integrated build is installed from PyPI, but without
constraints. This means the version installed may not be what is expected (as
we're currently seeing in the stable/pike branch).

This commit removes the tempest_git_* overrides from
playbooks/defaults/repo_packages/openstack_testing.yml, which results in
tempest getting installed from PyPI but with constraints applied.

Change-Id: I014201129b4d6b47ec45ff6239287623123a1595