Add default value for log_dirs in os-log-dir-setup.yml
There is no default value for log_dirs in the os-log-dir-setup.yml
which means that any role that includes it without that value will
crash and burn.  We set a default here to protect us from breaking
roles in the future if we forget to specifiy log_dirs on baremetal.

Change-Id: I07bf9f4bce7bb53ecc305d260142e878797d6c04
Closes-Bug: #1761781
(cherry picked from commit e9eca74e8366667fd0c26e47368e1092fee4590a)