Remove double warning from run-upgrade
The reality is that the upgrade script is always in
development, as is all open source software. Instead
of having this extra warning we just have one which
provides a suitable warning to use this in a test
environment first.

Also, we remove the set -v because the output of the
whole script is not very helpful when executing or
debugging.

Change-Id: I8e23bfacf131a0c7d5c3716437dbfecb896bbb70