Ensure that the repo servers also use the pypi cache
When re-running openstack-hosts after a greenfield deployment
the repo servers will be reconfigured to make use of the pypi
cache implemented on the repo host. There is therefore no
real reason why it should not be done sooner during the
greenfield deploy - otherwise the deployment is inconsistent.

Also, in a situation where the deployment has no containers,
not implementing the configuration to use the repo pypi cache
results in the rest of the deployment not being able to find
the packages which were built from source. It instead makes
all pip installs go to pypi and never even attempt the cache.

Finally, once the repo build is complete the repo-use playbook
is run to ensure that pip is reconfigured to use it - but the
facts to set the default pip index were set prior to the repo
being built, so the reconfiguration leaves the default pip
index to that of pypi. as such, in this patch, we re-verify
and re-set the applicable vars to ensure that the repo is used.

Depends-On: https://review.openstack.org/563106
Depends-On: https://review.openstack.org/563111
Change-Id: I2e7f9bc6cbf15b3309f843ffc24b4643563906a0
(cherry picked from commit 6677e1ef8f60ee64720288b748a19517a09f209d)