Ensure that repo checks use the right values
When executing the check for the appropriate files on
the repo host and then overriding the group_vars to
cater for when the repo is not present, we need to make
sure that we're doing the checks against the right URL.

In this patch we backup the default URL so that it can
be used when re-executing the check later in the same
playbook.

Also, given that the set-pip-upstream-url.yml and
set-upper-constraints.yml task sets are always used
together, we combine them into one task set.

Conflicts:
>------playbooks/containers-nspawn-create.yml

Change-Id: I040c4f5b6e96c85c68beb975f7b80d1edf42276d
(cherry picked from commit 792c753354381a030f11b949cd159c329c5cfdfe)