Ensure facts are updated for new containers
When the new containers are built, we currently only collect
a minimal amount of facts afterwards. This patch ensures that
a broader set of facts are updated to ensure that we have the
updated distribution information, networking and storage facts.

We do this for both types of containers to ensure consistency
in the approach.

Related-Bug: 1765438
Change-Id: I620e35d0942a6d4a55e05ae441d0e6ee2e4a67c4