Update role SHA's to resolve pip/python venv issues
With the release of pip10 and the use of python3 as the
default for the deploy host there were a bunch of patches
needed across all roles to ensure that the pinned versions
of pip/setuptools/wheel were actually being used for hosts,
containers and virtualenvs.

This patch updates the role SHA's to ensure that all these
fixes are brought in for the next release.

Closes-Bug: #1764470
Change-Id: I5552fe62a9567f43c9f7c958cc30fe7ddfc16d92