Adjust inside/outside openstack-ci bootstrap-host tasks
To optimise the execution and to simplify it, we re-arrange the
tasks in such a way that there are only two blocks - one for
execution inside openstack-infra, and one for outside.

Also, Zuul places a /etc/pip.conf file on the host but we do not
need it given that we have our own implementation of pip config
which uses the correct mirrors based on the mirror script.
As such, we no longer want this file in the containers, so we
remove the setting which makes lxc_hosts copy it into the
container cache when it is prepared.

Change-Id: I4986ff4d21e9fa8b77115ab80c66de99afe45f6c